feat(repo): 添加仓库卡片条目计数并优化数据类型
- 将 RepoCardVO 中的 likeCount 和 commentCount 类型从 Integer 改为 Long
- 在 RepoCardVO 中新增 itemCount 字段用于存储项目数量
- 更新 RepoConverter 映射配置以支持新的 itemCount 字段
- 修改 toCardVO 方法参数类型以匹配新的数据结构
- 将 RepoDetailVO 中的 likeCount 类型从 Integer 改为 Long
- 在 RepoItemService 接口中添加 count 方法用于统计项目数量
- 实现 RepoItemServiceImpl 中的 count 方法来查询仓库项目总数
- 根据仓库类型判断是否需要计算项目数量，代码类型仓库返回 0
- 更新仓库卡片转换逻辑以包含项目计数信息

+++ src/main/java/com/zincoid/me/model/vo/RepoDetailVO.java
@@ -25,7 +25,7 @@ public class RepoDetailVO {
     private String coverImage;
     private Boolean isDefaultCover;
     private Long viewCount;
-    private Integer likeCount;
+    private Long likeCount;
     private Boolean isLiked;
     private List<LikerVO> recentLikers;
     private Boolean restricted;
